only add the patch for initqueue --online option,keep other things untouch

+++ xCAT-server/share/xcat/netboot/sles/dracut_033/patch/base/initqueue.sh
@@ -46,6 +46,10 @@ exe=$1
 shift
 
 [ -x "$exe" ] || exe=$(command -v $exe)
+if [ -z "$exe" ] ; then
+    echo "Invalid command"
+    exit 1
+fi
 
 {
     [ -n "$onetime" ] && echo '[ -e "$job" ] && rm -f -- "$job"'
